refactor: Moved entry points to a separate file
Moved entry points and related functions from app.py to entry_points.py
