+++ main.py
@@ -1,7 +1,7 @@
 import sys
 sys.path.append("src")
-from unipoll_api import app  # noqa: E402
+from unipoll_api import entry_points  # noqa: E402
 
 
 if __name__ == "__main__":
-    app.cli_entry_point()
+    entry_points.cli_entry_point()
